Rename List to ComponentList
