+++ ui/src/KaptureApp/components/EditorSidebar/ComponentList/index.js
@@ -1,0 +1,1 @@
+export { ComponentList } from './ComponentList';
