+++ ui/src/KaptureApp/components/EditorSidebar/Components/Components.module.css
@@ -45,27 +45,3 @@
   padding: 0.8em;
   flex: 0 0 0%;
 }
-
-.list {
-  flex: 1;
-  overflow: auto;
-}
-
-.component {
-  cursor: pointer;
-  transition: all 0.1s ease;
-  padding: 0.2em 1em;
-  font-size: 0.8em;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.component:hover {
-  background-color: rgba(0, 0, 0, 0.03);
-}
-
-.typeCircle {
-  margin-right: 1em !important;
-}
